Movesw endpoint configuration to the settings menu
