+++ aas-web-ui/src/types/BaSyx.ts
@@ -3,6 +3,7 @@ import { Ref } from 'vue';
 export interface BaSyxComponent {
     url: Ref<string>;
     loading: Ref<boolean>;
+    connected: Ref<boolean | null>;
     connect: () => void;
     label: string;
     pathCheck?: string;
